Revert "retrofit and di integration "
