+++ app/src/main/AndroidManifest.xml
@@ -6,8 +6,6 @@
     </uses-permission>
 
     <application
-        android:usesCleartextTraffic="true"
-        android:name=".di.HiltApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
