--- app/src/main/java/com/techgeeksclub/earthquake/data/entity/EarthquakeResponse.kt
@@ -1,7 +1,0 @@
-package com.techgeeksclub.earthquake.data.entity
-
-import java.io.Serializable
-
-data class EarthquakeResponse(
-    var earthquakes : List<Earthquake>
-) : Serializable
