--- app/src/main/java/com/techgeeksclub/earthquake/di/HiltApplication.kt
@@ -1,8 +1,0 @@
-package com.techgeeksclub.earthquake.di
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class HiltApplication : Application(){
-}
