+++ app/src/main/java/com/techgeeksclub/earthquake/model/ClosestCities.kt
