+++ app/src/main/java/com/techgeeksclub/earthquake/model/ClosestCity.kt
