+++ app/src/main/java/com/techgeeksclub/earthquake/model/Coordinates.kt
