+++ app/src/main/java/com/techgeeksclub/earthquake/model/Earthquake.kt
