+++ app/src/main/java/com/techgeeksclub/earthquake/model/Geojson.kt
