+++ app/src/main/java/com/techgeeksclub/earthquake/model/LocationProperties.kt
