+++ app/src/main/java/com/techgeeksclub/earthquake/model/Metadata.kt
