+++ app/src/main/java/com/techgeeksclub/earthquake/model/Result.kt
