--- app/src/main/java/com/techgeeksclub/earthquake/retrofit/ApiUtils.kt
@@ -1,13 +1,0 @@
-package com.techgeeksclub.earthquake.retrofit
-
-class ApiUtils {
-
-    companion object{
-        val BASE_URL = "https://api.example.com/"
-
-
-        fun getEarthquakeDao() : EarthquakeDao{
-            return RetrofitClient.getClient(BASE_URL).create(EarthquakeDao::class.java)
-        }
-    }
-}
