--- app/src/main/java/com/techgeeksclub/earthquake/retrofit/EarthquakeDao.kt
@@ -1,10 +1,0 @@
-package com.techgeeksclub.earthquake.retrofit
-
-import com.techgeeksclub.earthquake.data.entity.EarthquakeResponse
-import retrofit2.http.GET
-
-interface EarthquakeDao {
-
-    @GET("earthquakes")
-    fun getEarthquakes(): EarthquakeResponse
-}
