--- app/src/main/java/com/techgeeksclub/earthquake/retrofit/RetrofitClient.kt
@@ -1,16 +1,0 @@
-package com.techgeeksclub.earthquake.retrofit
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitClient {
-
-    companion object {
-        fun getClient(baseUrl:String) : Retrofit {
-            return Retrofit.Builder()
-                .baseUrl(baseUrl)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-        }
-    }
-}
