+++ app/src/main/java/com/techgeeksclub/earthquake/ui/emergency/EmergencyViewModel.kt
@@ -1,0 +1,7 @@
+package com.techgeeksclub.earthquake.ui.emergency
+
+import androidx.lifecycle.ViewModel
+
+class EmergencyViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
