+++ app/src/main/java/com/techgeeksclub/earthquake/ui/home/HomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.techgeeksclub.earthquake.ui.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel() {
+
+}
