+++ app/src/main/java/com/techgeeksclub/earthquake/ui/settings/SettingsViewModel.kt
@@ -1,0 +1,7 @@
+package com.techgeeksclub.earthquake.ui.settings
+
+import androidx.lifecycle.ViewModel
+
+class SettingsViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
