--- app/src/main/java/com/techgeeksclub/earthquake/ui/viewmodel/HomeViewModel.kt
@@ -1,10 +1,0 @@
-package com.techgeeksclub.earthquake.ui.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class HomeViewModel @Inject constructor(): ViewModel() {
-
-}
