+++ app/src/main/res/layout/fragment_emergency.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.EmergencyFragment">
+    tools:context=".ui.emergency.EmergencyFragment">
 
     <TextView
         android:layout_width="wrap_content"
