+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.techgeeksclub.earthquake.ui.fragment.HomeFragment"
+        android:name="com.techgeeksclub.earthquake.ui.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" >
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="com.techgeeksclub.earthquake.ui.fragment.SettingsFragment"
+        android:name="com.techgeeksclub.earthquake.ui.settings.SettingsFragment"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" >
         <action
@@ -28,7 +28,7 @@
     </fragment>
     <fragment
         android:id="@+id/emergencyFragment"
-        android:name="com.techgeeksclub.earthquake.ui.fragment.EmergencyFragment"
+        android:name="com.techgeeksclub.earthquake.ui.emergency.EmergencyFragment"
         android:label="fragment_emergency"
         tools:layout="@layout/fragment_emergency" >
         <action
