+++ build.gradle
@@ -1,11 +1,4 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-
-buildscript {
-    dependencies {
-        classpath("androidx.navigation:navigation-safe-args-gradle-plugin:2.5.3")
-        classpath ("com.google.dagger:hilt-android-gradle-plugin:2.44")
-    }
-}
 plugins {
     id 'com.android.application' version '8.0.0' apply false
     id 'com.android.library' version '8.0.0' apply false
